DOP-2215: Move landing-domain components and tests
