+++ src/components/ComponentFactory.js
@@ -1,63 +1,70 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { ComponentFactory as LandingComponentFactory } from './landing';
-import Paragraph from './Paragraph';
+import Admonition, { admonitionMap } from './Admonition';
+import BlockQuote from './BlockQuote';
+import Button from './Button';
+import Card from './Card';
+import CardGroup from './CardGroup';
+import Code from './Code';
+import Cond from './Cond';
+import Container from './Container';
+import CSSClass from './CSSClass';
+import CTA from './CTA';
+import DefinitionList from './DefinitionList';
+import DefinitionListItem from './DefinitionListItem';
+import DeprecatedVersionSelector from './DeprecatedVersionSelector';
+import Describe from './Describe';
+import Emphasis from './Emphasis';
+import Extract from './Extract';
+import Field from './Field';
+import FieldList from './FieldList';
+import Figure from './Figure';
+import Footnote from './Footnote';
+import FootnoteReference from './FootnoteReference';
+import Glossary from './Glossary';
+import Heading from './Heading';
+import HorizontalList from './HorizontalList';
+import Image from './Image';
+import Include from './Include';
+import Introduction from './Introduction';
+import Kicker from './Kicker';
+import Line from './Line';
+import LineBlock from './LineBlock';
 import List from './List';
 import ListItem from './ListItem';
 import ListTable from './ListTable';
-import Emphasis from './Emphasis';
-import Include from './Include';
-import Section from './Section';
-import Code from './Code';
-import LiteralInclude from './LiteralInclude';
-import Tabs from './Tabs';
-import Admonition, { admonitionMap } from './Admonition';
-import Figure from './Figure';
 import Literal from './Literal';
-import Heading from './Heading';
-import BlockQuote from './BlockQuote';
-import Reference from './Reference';
-import Strong from './Strong';
-import URIWriter from './URIWriter/URIWriter';
-import TitleReference from './TitleReference';
-import Text from './Text';
-import DefinitionList from './DefinitionList';
-import DefinitionListItem from './DefinitionListItem';
-import Transition from './Transition';
-import CSSClass from './CSSClass';
-import SubstitutionReference from './SubstitutionReference';
-import Line from './Line';
-import LineBlock from './LineBlock';
-import HorizontalList from './HorizontalList';
-import Container from './Container';
-import Cond from './Cond';
-import Meta from './Meta';
-import VersionModified from './VersionModified';
-import CardGroup from './CardGroup';
-import Footnote from './Footnote';
-import FootnoteReference from './FootnoteReference';
 import LiteralBlock from './LiteralBlock';
-import Topic from './Topic';
-import Subscript from './Subscript';
-import Superscript from './Superscript';
-import Image from './Image';
+import LiteralInclude from './LiteralInclude';
+import Meta from './Meta';
+import MongoWebShell from './MongoWebShell';
+import OpenAPI from './OpenAPI';
+import Operation from './Operation';
+import Paragraph from './Paragraph';
+import Procedure from './Procedure';
+import Reference from './Reference';
 import RefRole from './RefRole';
-import Target from './Target';
-import Glossary from './Glossary';
+import ReleaseSpecification from './ReleaseSpecification';
+import Root from './Root';
 import Rubric from './Rubric';
 import SearchResults from './SearchResults';
-import Field from './Field';
-import FieldList from './FieldList';
-import Operation from './Operation';
-import OpenAPI from './OpenAPI';
-import Root from './Root';
-import StepYAML from './StepYAML';
+import Section from './Section';
+import Step from './Step';
 import StepsYAML from './StepsYAML';
-import MongoWebShell from './MongoWebShell';
-import Extract from './Extract';
-import Describe from './Describe';
-import ReleaseSpecification from './ReleaseSpecification';
+import StepYAML from './StepYAML';
+import Strong from './Strong';
+import Subscript from './Subscript';
+import SubstitutionReference from './SubstitutionReference';
+import Superscript from './Superscript';
+import Tabs from './Tabs';
+import Target from './Target';
+import Text from './Text';
+import TitleReference from './TitleReference';
+import Topic from './Topic';
+import Transition from './Transition';
 import Twitter from './Twitter';
+import URIWriter from './URIWriter/URIWriter';
+import VersionModified from './VersionModified';
 
 import RoleAbbr from './Roles/Abbr';
 import RoleClass from './Roles/Class';
@@ -111,15 +118,19 @@ const roleMap = {
 const componentMap = {
   admonition: Admonition,
   blockquote: BlockQuote,
+  button: Button,
+  card: Card,
   'card-group': CardGroup,
   class: CSSClass,
   code: Code,
   cond: Cond,
   container: Container,
   cssclass: CSSClass,
+  cta: CTA,
   definitionList: DefinitionList,
   definitionListItem: DefinitionListItem,
   deprecated: VersionModified,
+  'deprecated-version-selector': DeprecatedVersionSelector,
   describe: Describe,
   emphasis: Emphasis,
   extract: Extract,
@@ -133,6 +144,8 @@ const componentMap = {
   hlist: HorizontalList,
   image: Image,
   include: Include,
+  introduction: Introduction,
+  kicker: Kicker,
   line: Line,
   line_block: LineBlock,
   list: List,
@@ -147,6 +160,7 @@ const componentMap = {
   openapi: OpenAPI,
   operation: Operation,
   paragraph: Paragraph,
+  procedure: Procedure,
   ref_role: RefRole,
   reference: Reference,
   release_specification: ReleaseSpecification,
@@ -154,6 +168,7 @@ const componentMap = {
   rubric: Rubric,
   'search-results': SearchResults,
   section: Section,
+  step: Step,
   'step-yaml': StepYAML,
   'steps-yaml': StepsYAML,
   strong: Strong,
@@ -180,9 +195,8 @@ const ComponentFactory = (props) => {
       return null;
     }
 
-    if (domain === 'landing') {
-      return <LandingComponentFactory {...props} />;
-    }
+    // TODO: Create structure for domain handling?
+    console.log(domain);
 
     const lookup = type === 'directive' ? name : type;
     let ComponentType = componentMap[lookup];
