+++ src/components/Step.js
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { uiColors } from '@leafygreen-ui/palette';
-import ComponentFactory from '../ComponentFactory';
-import { theme } from '../../theme/docsTheme';
+import { theme } from '../theme/docsTheme';
+import ComponentFactory from './ComponentFactory';
 
 const Circle = styled('div')`
   background: ${uiColors.green.light3};
