--- src/components/landing/index.js
@@ -1,3 +1,0 @@
-import ComponentFactory from './ComponentFactory';
-
-export { ComponentFactory };
