+++ tests/unit/Button.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Button from '../../src/components/landing/Button';
+import Button from '../../src/components/Button';
 
 // data for this component
-import mockData from './data/landing/Button.test.json';
+import mockData from './data/Button.test.json';
 
 it('renders correctly', () => {
   const tree = shallow(<Button nodeData={mockData} />);
