+++ tests/unit/Procedure.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Procedure from '../../src/components/landing/Procedure';
+import Procedure from '../../src/components/Procedure';
 
 // data for this component
-import mockData from './data/landing/Procedure.test.json';
+import mockData from './data/Procedure.test.json';
 
 it('renders correctly', () => {
   const tree = shallow(<Procedure nodeData={mockData} />);
