+++ tests/unit/Step.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Step from '../../src/components/landing/Step';
+import Step from '../../src/components/Step';
 
 // data for this component
-import mockData from './data/landing/Step.test.json';
+import mockData from './data/Step.test.json';
 
 it('renders correctly', () => {
   const tree = shallow(<Step nodeData={mockData} />);
